Code clean up
Killed dead and redundant codes.

But there are some code that used the old key calculation routines is
hard to change. I've left FIXME marks on them. Please help to correct
them.

Reformatted according to coding style: src/kds_rpc_server.c

Changed horus_get_key() interface for better performance.

+++ lib/horus_crypt.c
@@ -26,7 +26,7 @@ horus_coding (int fd, int fdpos, char *buf, size_t nbyte,
   off_t offset;
   int ncopy;
   char key[SHA_DIGEST_LENGTH];
-  int key_len;
+  size_t key_len;
 
   rest = nbyte;
   bufp = 0;
